ai_assistant/requirements.txt
# Dependencias del Sistema de IA para SistemaAdmin
# Optimizado para plan gratuito de Render

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Base de datos
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# Procesamiento de datos
pandas==2.1.3
numpy==1.24.3

# Procesamiento de lenguaje natural
openai==1.3.7
transformers==4.35.2
torch==2.1.1
tokenizers==0.15.0

# Utilidades
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Logging y monitoreo
structlog==23.2.0

# Testing (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Seguridad
cryptography==41.0.7
